include gravity in flux, removing it from computations
