+++ external/vcm/vcm/__init__.py
@@ -47,7 +47,7 @@ from .sampling import train_test_split_sample
 from .derived_mapping import DerivedMapping
 from .cloud import get_fs
 from .cloud.fsspec import to_url
-from .projection import project_vertical_flux, flux_convergence
+from .projection import project_vertical_flux, flux_convergence, cell_center_convergence
 
 __all__ = [item for item in dir() if not item.startswith("_")]
 __version__ = "0.1.0"
